add auth to server calls

+++ lib/project2_web/controllers/reply_controller.ex
@@ -6,12 +6,16 @@ defmodule Project2Web.ReplyController do
 
   action_fallback Project2Web.FallbackController
 
+  plug Project2Web.Plugs.RequireAuth when action in [:create, :update, :delete]
+  
   def index(conn, _params) do
     replies = Replies.list_replies()
     render(conn, "index.json", replies: replies)
   end
 
   def create(conn, %{"reply" => reply_params}) do
+    reply_params = Map.put(reply_params, "user_id", conn.assigns.current_user.id)
+    
     with {:ok, %Reply{} = reply} <- Replies.create_reply(reply_params) do
       conn
       |> put_status(:created)
@@ -36,8 +40,14 @@ defmodule Project2Web.ReplyController do
   def delete(conn, %{"id" => id}) do
     reply = Replies.get_reply!(id)
 
-    with {:ok, %Reply{}} <- Replies.delete_reply(reply) do
-      send_resp(conn, :no_content, "")
+    if reply.user_id == conn.assigns.current_user.id do 
+      with {:ok, %Reply{}} <- Replies.delete_reply(reply) do
+        send_resp(conn, :no_content, "")
+      end
+    else 
+      conn 
+      |> put_resp_header("content-type", "application/json; charset=UTF-8")
+      |> send_resp(:unprocessable_entity, Jason.encode!(%{"error" => "not authorized!"}) )
     end
   end
 end
